Moved prototypes for process_normal and process_win32 to include/common.h to avoid compiler warnings and repetitions.
Cleaned up audit code slightly

Fixed display banner to treat current working directory
as a Unicode string.

Enabled and allowed the user load known files that have
different kinds of hashes in them.

Version bump to 3.0. I've made so many changes to the
common code that it warrants the jump.